+++ components/Fridge/FridgeDetail.js
