--- components/Map/styles.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export const MapWrapper = styled.div`
-  overflow: hidden;
-  width: calc(100vw - 591px);
-  height: calc(100vh - 137px);
-`;
